backend/requirements.txt
alembic==1.4.3
astroid==2.4.2
Authlib==0.15.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
ecdsa==0.16.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
future==0.18.2
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.6
pycparser==2.20
pycryptodome==3.3.1
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose-cryptodome==1.3.2
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
